Update: ProductItem 컴포넌트 분리 추가
