+++ src/components/ProductItem/index.ts
@@ -1,0 +1,1 @@
+export { default as ProductItem } from './ProductItem';
